Addressing Pull Request Feedback for Description Annotations.
